Split ScratchItem* from ScratchList
