+++ frontend/src/components/user/tabs/ScratchesTab.tsx
@@ -1,4 +1,6 @@
-import ScratchList, { ScratchItemNoOwner } from "@/components/ScratchList";
+import ScratchList from "@/components/ScratchList";
+import { ScratchItemNoOwner } from "@/components/ScratchItem";
+
 import type { User } from "@/lib/api";
 import { userUrl } from "@/lib/api/urls";
